refactor: restructure layout domain organization
- Remove intermediate components directory in layout domain
- Move DesignSystemSection and TechnicalImplementationSection directly under layout/
- Rename Layout.tsx to component.tsx and Layout.module.css to component.module.css
- Move skillExtraction.ts to scripts/ directory for better organization
- Update all import paths to reflect new flat structure
