+++ src/domains/layout/DesignSystemSection/DesignSystemSection.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Palette, ExternalLink, Github } from 'lucide-react';
 import styles from './DesignSystemSection.module.css';
-import layoutStyles from '../Layout/Layout.module.css';
+import layoutStyles from '../component.module.css';
 
 export const DesignSystemSection: React.FC = () => {
   return (
