+++ src/domains/layout/TechnicalImplementationSection/TechnicalImplementationSection.module.css
