+++ src/domains/layout/TechnicalImplementationSection/TechnicalImplementationSection.tsx
