+++ src/domains/layout/component.tsx
@@ -1,13 +1,13 @@
 import React from 'react';
 import { ApolloProvider } from '@apollo/client';
 import { Code } from 'lucide-react';
-import { client } from '../../../../apollo/client';
-import { RecommendationsSection } from '../../../recommendations/components/RecommendationsSection';
-import { usePageTracking } from '../../../analytics';
-import styles from './Layout.module.css';
-import { DisplayFlex, DisplayFlexItem } from '../../../../shared/components/ui/DisplayFlex';
-import { DesignSystemSection } from '../DesignSystemSection';
-import { TechnicalImplementationSection } from '../TechnicalImplementationSection/TechnicalImplementationSection';
+import { client } from '../../apollo/client';
+import { RecommendationsSection } from '../recommendations/components/RecommendationsSection';
+import { usePageTracking } from '../analytics';
+import styles from './component.module.css';
+import { DisplayFlex, DisplayFlexItem } from '../../shared/components/ui/DisplayFlex';
+import { DesignSystemSection } from './DesignSystemSection';
+import { TechnicalImplementationSection } from './TechnicalImplementationSection/TechnicalImplementationSection';
 
 export const Layout: React.FC = () => {
   // Track page views and scroll behavior
